requirements.txt
streamlit>=1.30
tensorflow==2.16.1
numpy
pillow
